second nos case added

+++ SmICSCoreLib/Factories/NUMNode/NUMNodeModel.cs
@@ -6,13 +6,13 @@ namespace SmICSCoreLib.Factories.NUMNode
     public class NUMNodeModel
     {
         [JsonProperty(PropertyName = "AverageNumberOfStays")]
-        public int AverageNumberOfStays { get; set; }
+        public double AverageNumberOfStays { get; set; }
         [JsonProperty(PropertyName = "AverageNumberOfNosCases")]
-        public int AverageNumberOfNosCases { get; set; }
+        public double AverageNumberOfNosCases { get; set; }
         [JsonProperty(PropertyName = "AverageNumberOfMaybeNosCases")]
-        public int AverageNumberOfMaybeNosCases { get; set; }
+        public double AverageNumberOfMaybeNosCases { get; set; }
         [JsonProperty(PropertyName = "AverageNumberOfContacts")]
-        public int AverageNumberOfContacts { get; set; }
+        public double AverageNumberOfContacts { get; set; }
         [JsonProperty(PropertyName = "DateTime")]
         public DateTime DateTime { get; set; }
